feat: add auth to to-do resolvers

+++ src/todo/todo.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateTodoInput, UpdateTodoInput, PaginationArgs } from './dto';
 import { UserInputError } from '@nestjs/apollo';
@@ -6,30 +10,41 @@ import { UserInputError } from '@nestjs/apollo';
 @Injectable()
 export class TodoService {
   constructor(private prisma: PrismaService) {}
-  create(createTodoInput: CreateTodoInput) {
+
+  create(createTodoInput: CreateTodoInput, context: any) {
     if (!createTodoInput.title)
-      throw new Error('Please provide a title for your to-do!');
-    return this.prisma.todo.create({ data: { ...createTodoInput } });
+      throw new BadRequestException('Please provide a title for your to-do!');
+    return this.prisma.todo.create({
+      data: { ...createTodoInput, user: { connect: { id: context.id } } },
+    });
   }
 
-  findAll(paginationArgs?: PaginationArgs) {
-    return this.prisma.todo.findMany(paginationArgs);
+  findAll(paginationArgs: PaginationArgs, context: any) {
+    return this.prisma.todo.findMany({
+      where: { userId: context.id },
+      ...paginationArgs,
+    });
   }
 
-  async findOne(id: number) {
+  async findOne(id: number, context: any) {
     if (!id || id <= 0)
       throw new UserInputError('Please provide a valid ID to find a to-do.');
-    const todo = await this.prisma.todo.findUnique({ where: { id } });
-    if (!todo) throw new Error('No to-do found with the specified ID.');
+    const todo = await this.prisma.todo.findFirst({
+      where: { id, userId: context.id },
+    });
+    if (!todo) {
+      throw new NotFoundException('No to-do found with the specified ID.');
+    }
     return todo;
   }
 
-  async searchTitleDesc(searchQuery: string) {
+  async searchTitleDesc(searchQuery: string, context: any) {
     if (!searchQuery)
       throw new UserInputError('Please provide a search query.');
     if (!searchQuery.trim())
       throw new UserInputError('Please provide a non-empty search query.');
-    return await this.prisma.todo.findMany({
+
+    const todoMatches = await this.prisma.todo.findMany({
       where: {
         OR: [
           {
@@ -43,11 +58,13 @@ export class TodoService {
             },
           },
         ],
+        userId: context.id,
       },
     });
+    return todoMatches;
   }
 
-  async update(id: number, updateTodoInput: UpdateTodoInput) {
+  async update(id: number, updateTodoInput: UpdateTodoInput, context: any) {
     if (!id || id <= 0)
       throw new UserInputError('Please provide a valid ID to update a to-do.');
     if (!updateTodoInput)
@@ -56,17 +73,21 @@ export class TodoService {
       );
     const todo = await this.prisma.todo.findUnique({ where: { id } });
     if (!todo) throw new Error('No to-do found with the specified ID.');
+    if (todo.userId !== context.id)
+      throw new Error("You can't edit a to-do that does not belong to you");
     return this.prisma.todo.update({
       where: { id },
       data: { ...updateTodoInput },
     });
   }
 
-  async remove(id: number) {
+  async remove(id: number, context: any) {
     if (!id || id <= 0)
       throw new UserInputError('Please provide a valid ID to remove a to-do.');
     const todo = await this.prisma.todo.findUnique({ where: { id } });
     if (!todo) throw new Error('No to-do found with the specified ID.');
+    if (todo.userId !== context.id)
+      throw new Error("You can't delete a to-do that does not belong to you");
     return this.prisma.todo.delete({ where: { id } });
   }
 }
